+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Changelog
 
-## [0.1.0+8] - 2026-04-26
+
+## [Unreleased]
+
+
+## [0.1.0+9] - 2026-04-23
+
+### Fixed
+- **Settings notifier** now works immediatelly
+- **Adjustment display panel** sizing and placing, text size clamp
+- **Conditions screen icon**
+- **Reticle view display ratio**
+
+### Changed
+- **Max window size** limitations are disabled 
+
+### Features
+- **Improve home screen** condition indicators got text labels 
+- **Improve home paging** pages got text labels 
+- **Display adjustments in current clicks** on home screen, tables screen, html report and reticle view screen
+
+
+## [0.1.0+8] - 2026-04-23
 
 ### Fixed
 - **Hotfix:** corrected twist rate validation — the field now accepts `0` as a valid value.